Code motion and renaming related to the bulk file move that created the `Backend` module tree.
The main change here is that the `Cryptol.Eval.SBV` module was split
into two, similar to how the concrete and What4 evaluators were already
split.  Various other small bits of code are rearranged to detangle
module dependencies.
