+++ cryptol.cabal
@@ -160,24 +160,24 @@ library
 
                        Cryptol.IR.FreeVars,
 
+                       Cryptol.Backend,
+                       Cryptol.Backend.Concrete,
+                       Cryptol.Backend.FloatHelpers,
+                       Cryptol.Backend.Monad,
+                       Cryptol.Backend.SBV,
+                       Cryptol.Backend.What4,
+                       Cryptol.Backend.What4.SFloat,
+
                        Cryptol.Eval,
                        Cryptol.Eval.Arch,
-                       Cryptol.Eval.Backend,
                        Cryptol.Eval.Concrete,
-                       Cryptol.Eval.Concrete.Float,
-                       Cryptol.Eval.Concrete.FloatHelpers,
-                       Cryptol.Eval.Concrete.Value,
                        Cryptol.Eval.Env,
                        Cryptol.Eval.Generic,
-                       Cryptol.Eval.Monad,
                        Cryptol.Eval.Reference,
                        Cryptol.Eval.SBV,
                        Cryptol.Eval.Type,
                        Cryptol.Eval.Value,
                        Cryptol.Eval.What4,
-                       Cryptol.Eval.What4.Value,
-                       Cryptol.Eval.What4.Float,
-                       Cryptol.Eval.What4.SFloat,
 
                        Cryptol.AES,
                        Cryptol.SHA,
