+++ src/Cryptol/Eval/Env.hs
@@ -14,8 +14,10 @@
 {-# LANGUAGE DeriveGeneric #-}
 module Cryptol.Eval.Env where
 
-import Cryptol.Eval.Backend
-import Cryptol.Eval.Monad( PPOpts )
+import Cryptol.Backend
+
+import Cryptol.Backend.Monad( PPOpts )
+
 import Cryptol.Eval.Type
 import Cryptol.Eval.Value
 import Cryptol.ModuleSystem.Name
