+++ src/Cryptol/REPL/Monad.hs
@@ -96,7 +96,7 @@ import qualified Cryptol.Parser.AST as P
 import Cryptol.Symbolic (SatNum(..))
 import Cryptol.Symbolic.SBV (SBVPortfolioException)
 import Cryptol.Symbolic.What4 (W4Exception)
-import Cryptol.Eval.Monad(PPFloatFormat(..),PPFloatExp(..))
+import Cryptol.Backend.Monad(PPFloatFormat(..),PPFloatExp(..))
 import qualified Cryptol.Symbolic.SBV as SBV (proverNames, setupProver, defaultProver, SBVProverConfig)
 import qualified Cryptol.Symbolic.What4 as W4 (proverNames, setupProver, W4ProverConfig)
